File uploads via uploadthing (#14)
Add file upload support via upload thing. This required using their
"custom adapter" docs for the backend, and setting up a vite proxy to
get my auth for free.... Probably should have been doing this for a
while but here we are!

Oh, and I guess we got toasts working too!

This is largely unfinished, we still need to actually do something on
the data side with these files. This at least gets everything ready for
that work.